backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
aiofiles==24.1.0
anthropic
openai
qdrant-client==1.12.0
pydantic>=2.7,<3
pydantic-settings>=2.3,<3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openpyxl==3.1.5
reportlab==4.2.5
